Need to differentiate between cached dns response vs incomplete ones (#400)
* Need to differentiate between cached dns response vs incomplete ones


* drop go1.21 CI since its running g1.22


---------

+++ pkg/decode/decode_protobuf_test.go
@@ -101,7 +101,7 @@ func TestPBFlowToMap(t *testing.T) {
 		"DnsId":                  uint16(1),
 		"DnsFlags":               uint16(0x80),
 		"DnsFlagsResponseCode":   "NoError",
-		"DnsErrno":               uint32(0),
+		"DnsErrno":               uint8(0),
 		"TimeFlowRttNs":          someDuration.Nanoseconds(),
 	}, out)
